pmlock & pmcd state fixups
- man page changes for pmlock -i ident option
- qa/1441 tweaks (esp. for older versions of hostnamectl(1))
- fix compiler warning in src/pmcd/src/dofetch.c

+++ src/pmcd/src/dofetch.c
@@ -641,7 +641,7 @@ HandleFetch(ClientInfo *cip, __pmPDU* pb, int pdutype)
     if (cip->status.changes) {
 	/* notify client of PMCD state change */
 	if (pmDebugOptions.appl6) {
-	    fprintf(stderr, "HandleFetch: client[%d] (fd %d) sent ", cip - client, cip->fd);
+	    fprintf(stderr, "HandleFetch: client[%d] (fd %d) sent ", (int)(cip - client), cip->fd);
 	    __pmDumpFetchFlags(stderr, cip->status.changes);
 	    fputc('\n', stderr);
 	}
